Create Scalasca_ScoreP_Cube4
This instruction is for installing Scalasca, Score-P, and Cube4 in Ubuntu 18.04LTS. Please check the links provided for other OSs. 

Score-P is the instrument tool and can be run with/without Scalasca. Scalasca is an integrating command for instrumenting and analyzing the results. To view the results, Cube4 has a GUI as a visualization tool. Another better visualization tool to install is Vampir, but I have not tried it.