Create Perf
Perf profiling is the easiest, compared to Darshan, ScoreP_Scalasca because it's used in Ubuntu normally to profile the Ubuntu performance itself.  Perf is good for knowing the overhead created by the code we want to profile. But the results will not show the performance bottleneck if we want to check parallel (MPI) program, or even for IO writing. Have a look at Darshan & Scalasca for MPI.